requirements.txt
discord.py
pytz
aiohttp
python-dotenv